requirements.txt
nicegui>=1.3.16
torch>=2.0.0
torchaudio>=2.0.0
demucs>=4.0.0
pyannote.audio>=3.1.0
pydub>=0.25.1
numpy>=1.24.0
pandas>=2.0.0
soundfile>=0.12.1
matplotlib>=3.7.0
python-dotenv>=1.2.2
